+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+### Added
+
+### Fixed
+- fix salutation in order an customer
+
+### Changed
+
+
 ## [4.3.0]
 ### Added
 - fallback for the import of the weight (@jppeter)